HW 89 added delete buttons and publish toggle

+++ api/routers/albums.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import { imagesUpload } from "../multer";
 import Album from "../models/Album";
-import auth, {RequestWithUser} from "../middleware/auth";
+import auth, { RequestWithUser } from "../middleware/auth";
 import permit from "../middleware/permit";
 import { HydratedDocument } from "mongoose";
 import { AlbumData } from "../types";
@@ -11,12 +11,14 @@ import role from "../middleware/role";
 const albumRouter = express.Router();
 
 albumRouter.post("/", auth, imagesUpload.single("photo"), async (req, res) => {
+  const user = (req as RequestWithUser).user;
   try {
     const album = await Album.create({
       name: req.body.name,
       artist: req.body.artist,
       photo: req.file ? req.file.filename : null,
       year: req.body.year,
+      author: user._id,
     });
     return res.send(album);
   } catch (error) {
@@ -24,28 +26,32 @@ albumRouter.post("/", auth, imagesUpload.single("photo"), async (req, res) => {
   }
 });
 
-albumRouter.get("/",role, async (req, res) => {
+albumRouter.get("/", role, async (req, res) => {
   try {
     const queryArtist = req.query.artist as string;
     const user = (req as RequestWithUser).user;
     if (queryArtist) {
-      if (user && user.role  === "admin") {
+      if (user && user.role === "admin") {
         const albums = await Album.find({
           artist: queryArtist,
         })
-            .populate("artist", "name")
-            .sort({ year: -1 });
+          .populate("artist", "name")
+          .sort({ year: -1 });
         return res.send(albums);
       }
       const albums = await Album.find({
-        isPublished: true,
-        artist: queryArtist,
+        $or: [
+          { isPublished: true, artist: queryArtist },
+          { author: user?._id, artist: queryArtist },
+        ],
       })
         .populate("artist", "name")
         .sort({ year: -1 });
       return res.send(albums);
     } else {
-      const albums = await Album.find({ isPublished: true })
+      const albums = await Album.find({
+        $or: [{ isPublished: true }, { author: user?._id }],
+      })
         .populate("artist", "name")
         .sort({ year: -1 });
       return res.send(albums);
@@ -87,11 +93,29 @@ albumRouter.patch(
   }
 );
 
-albumRouter.delete("/:id", auth, permit("admin"), async (req, res) => {
+albumRouter.delete("/:id", auth, async (req, res) => {
+  const user = (req as RequestWithUser).user;
   try {
-    await Track.deleteMany({ album: req.params.id });
-    await Album.deleteOne({ _id: req.params.id });
-    return res.send({ message: "deleted" });
+    if (user.role === "admin") {
+      const deleted = await Album.deleteOne({ _id: req.params.id });
+      if (deleted.deletedCount === 1) {
+        await Track.deleteMany({ album: req.params.id });
+        return res.send({ message: "deleted" });
+      } else {
+        return res.send({ message: "cant find album" });
+      }
+    }
+    const deleted = await Album.deleteOne({
+      _id: req.params.id,
+      author: user._id,
+      isPublished: false,
+    });
+    if (deleted.deletedCount === 1) {
+      await Track.deleteMany({ album: req.params.id });
+      return res.send({ message: "deleted" });
+    } else {
+      return res.send({ message: "no permission" });
+    }
   } catch {
     return res.sendStatus(500);
   }
